fold scheduler sql persistence into core persistence modules
- MySQL/Postgres/SqliteSchedulerConfiguration DELETED; scheduler DAO beans move onto
  MySQLConfiguration/PostgresConfiguration/SqliteConfiguration, gated conductor.scheduler.enabled=true
- scheduler/{mysql,postgres,sqlite}-persistence modules removed (settings.gradle, server deps, ci filters)
- scheduler DAO classes + tests moved into core persistence modules (packages unchanged)
- smoke fixture targets core configs; active cases run the real core flyway chain
- fixture cleanups setAutoCommit(true) — core module test datasources run hikari auto-commit=false
- known: mysql/postgres GrpcEndToEndTest component-scan collision w/ moved @TestConfiguration
  classes; grpc scan fix owned separately
