+++ scheduler/core/src/testFixtures/java/io/orkes/conductor/scheduler/service/AbstractSchedulerServiceIntegrationTest.java
@@ -69,6 +69,8 @@ public abstract class AbstractSchedulerServiceIntegrationTest {
     @Before
     public final void setUpService() throws Exception {
         try (Connection conn = dataSource().getConnection()) {
+            // module test datasources may set hikari auto-commit=false — deletes must commit
+            conn.setAutoCommit(true);
             conn.prepareStatement("DELETE FROM scheduler_execution").executeUpdate();
             conn.prepareStatement("DELETE FROM scheduler").executeUpdate();
             conn.prepareStatement("DELETE FROM workflow_scheduled_executions").executeUpdate();
