--- scheduler/mysql-persistence/build.gradle
@@ -1,29 +1,0 @@
-dependencies {
-    implementation project(':conductor-scheduler-core')
-    implementation project(':conductor-common')
-    implementation project(':conductor-core')
-    implementation project(':conductor-mysql-persistence')
-
-    compileOnly 'org.springframework.boot:spring-boot-starter'
-    compileOnly 'org.springframework.retry:spring-retry'
-
-    implementation "com.fasterxml.jackson.core:jackson-databind"
-    implementation "com.fasterxml.jackson.core:jackson-core"
-
-    implementation "org.apache.commons:commons-lang3"
-    implementation "mysql:mysql-connector-java:8.0.33"
-    implementation "org.flywaydb:flyway-core:${revFlyway}"
-    implementation "org.flywaydb:flyway-mysql:${revFlyway}"
-
-    testImplementation 'org.springframework.boot:spring-boot-starter'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.springframework.retry:spring-retry'
-    testImplementation "org.testcontainers:mysql:${revTestContainer}"
-    testImplementation testFixtures(project(':conductor-scheduler-core'))
-}
-
-test {
-    maxParallelForks = 1
-    // Docker Desktop on macOS requires API v1.44+; docker-java defaults to v1.41
-    environment 'DOCKER_API_VERSION', '1.44'
-}
