--- scheduler/postgres-persistence/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-org.conductoross.conductor.scheduler.postgres.config.PostgresSchedulerConfiguration
