--- scheduler/postgres-persistence/src/test/resources/application.properties
@@ -1,2 +1,0 @@
-conductor.scheduler.enabled=true
-spring.flyway.enabled=false
