--- scheduler/sqlite-persistence/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-org.conductoross.conductor.scheduler.sqlite.config.SqliteSchedulerConfiguration
