+++ sqlite-persistence/build.gradle
@@ -2,6 +2,7 @@ dependencies {
     implementation project(':conductor-common-persistence')
     implementation project(':conductor-common')
     implementation project(':conductor-core')
+    implementation project(':conductor-scheduler-core')
 
     compileOnly 'org.springframework.boot:spring-boot-starter'
     compileOnly 'org.springframework.retry:spring-retry'
@@ -29,6 +30,7 @@ dependencies {
 
     testImplementation project(':conductor-test-util').sourceSets.test.output
     testImplementation project(':conductor-common-persistence').sourceSets.test.output
+    testImplementation testFixtures(project(':conductor-scheduler-core'))
     testImplementation "org.awaitility:awaitility:${revAwaitility}"
 }
